% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{model_predict}
\alias{model_predict}
\title{model_predict}
\usage{
model_predict(model, newx, s = NA_real_)
}
\arguments{
\item{model}{model}

\item{newx}{newx}

\item{s}{default NAN.}
}
\description{
model_predict
}
